@@ -1,67 +1,71 @@
 package net.simplycrafted.DonationTracker;
 
+import lombok.Getter;
+import org.bukkit.command.CommandExecutor;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitScheduler;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.TreeMap;
 
 /**
  * Copyright © Brian Ronald
  * 28/06/14
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  */
 public class DonationTracker extends JavaPlugin {
-
+    
     // Our only instance
     private static DonationTracker donationtracker;
-
+    
     // List of instantiated goals
-    Map<String,Goal> goals;
-    Map<String,Goal> goalsBackwards;
-
+    @Getter
+    private Map<String, Goal> goals;
+    @Getter
+    private Map<String, Goal> goalsBackwards;
+    
     // Determine which goals need to be rewarded or otherwise
-    public void assess(boolean atDonationTime) {
-        for(String key: goals.keySet())
-        {
-            Goal goal = goals.get(key);
-            if (goal.reached()) {
-                if (atDonationTime) {
+    public void assess(final boolean atDonationTime) {
+        for(final Entry<String, Goal> stringGoalEntry : goals.entrySet()) {
+            final Goal goal = stringGoalEntry.getValue();
+            if(goal.reached()) {
+                if(atDonationTime) {
                     goal.ondonate();
                 }
                 goal.enable();
             }
         }
-        for(String key: goalsBackwards.keySet())
-        {
-            Goal goal = goalsBackwards.get(key);
-            if (!goal.reached()) {
+        for(final Entry<String, Goal> stringGoalEntry : goalsBackwards.entrySet()) {
+            final Goal goal = stringGoalEntry.getValue();
+            if(!goal.reached()) {
                 goal.abandon();
             }
         }
     }
-
+    
     // Withdraw all rewards (used when closing down the plugin).
     public void withdraw() {
-        for(String key: goalsBackwards.keySet())
-        {
-            Goal goal = goalsBackwards.get(key);
-            if (goal.reached()) {
+        for(final Entry<String, Goal> stringGoalEntry : goalsBackwards.entrySet()) {
+            final Goal goal = stringGoalEntry.getValue();
+            if(goal.reached()) {
                 getLogger().info("Abandoning " + goal.getName());
                 goal.abandon();
             }
         }
     }
-
+    
     @Override
     public void onEnable() {
         saveDefaultConfig();
@@ -69,48 +73,45 @@ public class DonationTracker extends JavaPlugin {
         goals = new TreeMap<>();
         goalsBackwards = new TreeMap<>(Collections.reverseOrder());
         // Bail out now if we have no database
-        Database dbtest = new Database();
-        if (dbtest.connectionIsDead()) return;
-        CommandHandler commandHandler = new CommandHandler();
+        final Database dbtest = new Database();
+        if(dbtest.connectionIsDead()) {
+            return;
+        }
+        final CommandExecutor commandHandler = new CommandHandler();
         getCommand("donation").setExecutor(commandHandler);
         getCommand("donorgoal").setExecutor(commandHandler);
         getCommand("donationpool").setExecutor(commandHandler);
         getCommand("dgadmin").setExecutor(commandHandler);
-
+        
         // Load the goals from the config file
         ConfigurationSection goalConfig;
-        ConfigurationSection goalsConfig = getConfig().getConfigurationSection("goals");
-        for (String key : goalsConfig.getKeys(false)) {
+        final ConfigurationSection goalsConfig = getConfig().getConfigurationSection("goals");
+        for(final String key : goalsConfig.getKeys(false)) {
             goalConfig = goalsConfig.getConfigurationSection(key);
-            Goal goal = new Goal (goalConfig);
-            if (goal.reached()) {
+            final Goal goal = new Goal(goalConfig);
+            if(goal.reached()) {
                 getLogger().info("...reached");
             }
             goals.put(key, goal);
             goalsBackwards.put(key, goal);
         }
-
+        
         // Schedule a checker to examine these goals periodically
-        BukkitScheduler scheduler = getServer().getScheduler();
-        scheduler.scheduleSyncRepeatingTask(this,new Runnable() {
-            @Override
-            public void run() {
-                assess(false);
-            }
-        },100,getConfig().getInt("period"));
+        final BukkitScheduler scheduler = getServer().getScheduler();
+        scheduler.scheduleSyncRepeatingTask(this, () -> assess(false), 100, getConfig().getInt("period"));
     }
-
+    
     @Override
     public void onDisable() {
         withdraw();
         Database.disconnect();
     }
-
+    
     // Lets other classes get a reference to our instance
     public static DonationTracker getInstance() {
         return donationtracker;
     }
-
+    
     public void reload() {
         // Withdraw all goal rewards
         withdraw();
@@ -121,11 +122,11 @@ public class DonationTracker extends JavaPlugin {
         reloadConfig();
         // Re-populate the goal hashes
         ConfigurationSection goalConfig;
-        ConfigurationSection goalsConfig = getConfig().getConfigurationSection("goals");
-        for (String key : goalsConfig.getKeys(false)) {
+        final ConfigurationSection goalsConfig = getConfig().getConfigurationSection("goals");
+        for(final String key : goalsConfig.getKeys(false)) {
             goalConfig = goalsConfig.getConfigurationSection(key);
-            Goal goal = new Goal (goalConfig);
-            if (goal.reached()) {
+            final Goal goal = new Goal(goalConfig);
+            if(goal.reached()) {
                 getLogger().info("...reached");
             }
             goals.put(key, goal);
